Safari-609.1.11.1/LayoutTests/plugins/netscape-plugin-setwindow-size-expected.txt

CONSOLE MESSAGE: PLUGIN: NPP_SetWindow: 800 200

This tests that a plug-in with a percentage width gets a correct NPP_SetWindow the first time.
